xcb: support nested paint operations in remote scenarios
Recent changes to the xcb platform plugin added logic to paint to a
server-side X Pixmap when the MIT-SHM extension is unavailable. Those
changes also added logic to keep track of the dirty areas that need to
be flushed to the server from the backing store when the backing store
is flushed to a window.

Because a single QRegion was used to track those areas, nested paint
operations no longer accurately tracked all the dirty areas. Consider:

void Widget::paintEvent(QPaintEvent *) {
    QPainter p1(this);

    for (int i = 0; longRunningOperationInProgress(); i++) {
        QPixmap pixmap(size());
        QPainter p2(&pixmap);
        p2.drawText(pixmap.rect(), i & 1 ? "Working." : "Working..");
        p2.end();
        QBackingStore *bs = backingStore();
        bs->beginPaint(rect());
        p1.drawPixmap(0, 0, pixmap);
        bs->endPaint();
        bs->flush(rect(), windowHandle());
    }

    p1.fillRect(rect(), Qt::red);
    p1.drawText(rect(), "Done!");
}

While this code could be restructured, it is a common pattern in legacy
Qt applications which cannot be easily changed.

Change-Id: I3b919266abe41c96a584cb02f41cafac3f9d1d7c
Reviewed-by: Laszlo Agocs <laszlo.agocs@theqtcompany.com>