Cocoa: Allow pasting text from Qt apps to Mail
Apple Mail will pick up and attempt to use the vCard
flavor, which then fails since the data Qt placed
on the clipboard is not actually a valid vCard.

Place data in the vCard format on the clipboard
only if the mime type is “text/vcard”.

[ChangeLog][OS X] Pasting text from Qt applications
to Apple Mail now works.

[ChangeLog][OS X] “text/vcard” is now required as
the mime type when placing vCards on the clipboard.

Task-number: QTBUG-48953
Change-Id: Id029b20317f2c5ad8ae225912484de3c97498d29
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@theqtcompany.com>
Reviewed-by: Jake Petroules <jake.petroules@qt.io>
Reviewed-by: Gabriel de Dietrich <gabriel.dedietrich@qt.io>