Applying widget toolTipDuration to items
If you set toolTipDuration to QListView widget,
the toolTip at the blank area is displayed exactly for the specified duration and then disppear.
But toolTip at item is displayed forever. It should be displayed for the same duration.
This pull request fixes that problem by applying the widget toopTipDuration to item toolTip.